// packages/frontend/src/generated/types/BlendingResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { DocumentStatus } from './DocumentStatus.ts'

/**
 * @description Response DTO for the `blending_document` entity.
 */
export interface BlendingResponse {
  /**
   * @type string, uuid
   */
  contractorId: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  date: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string,null
   */
  executedAt?: string | null
  /**
   * @type string,null, uuid
   */
  executedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string,null
   */
  revertedAt?: string | null
  /**
   * @type string,null, uuid
   */
  revertedBy?: string | null
  /**
   * @type string
   */
  status: DocumentStatus
  /**
   * @type string, uuid
   */
  targetProductId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}